pyproject.toml
[project]
name = "second-brain"
version = "0.1.0"
description = "Multi-agent second brain system"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic-ai",
    "pydantic-evals",
    "chromadb",
    "anthropic",
    "mem0ai",
    "spacy",
    "opentelemetry-api",
    "opentelemetry-sdk",
    "pypdf",
    "sentence-transformers",
    "opentelemetry-exporter-otlp-proto-grpc>=1.38.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
